+++ package.json
@@ -54,7 +54,7 @@
     "dayjs": "^1.11.10",
     "dedent": "^1.5.3",
     "lucide-react": "^0.456.0",
-    "next": "^15.0.2",
+    "next": "^15.0.3",
     "radix-ui": "^1.0.1",
     "react": "^19.0.0-rc-0bc30748-20241028",
     "react-dom": "^19.0.0-rc-0bc30748-20241028",
